% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{calculate_textplex}
\alias{calculate_textplex}
\title{Calculate the raw scores}
\usage{
calculate_textplex(input_text, udpipe_model, mtld_threshold = 0.72,
  mtld_min_ntoken = 10)
}
\arguments{
\item{input_text}{vector of characters.}

\item{udpipe_model}{udpipe_model loaded with udpipe::udpipe_load_model().}

\item{mtld_threshold}{Threshold for the calculation of MTLD score.}

\item{mtld_min_ntoken}{Threhold for minimal length of a viable sentence.}
}
\value{
tibble of raw scores
}
\description{
This function takes a vector of characters and calculate the raw scores required for the algorithm.
}
\examples{
\dontrun{
require(udpipe)
data(fairy)
udpipe_model <- udpipe_load_model('english-ewt-ud-2.4-190531.udpipe')
fairy_rawscore <- calculate_textplex(fairy$text, udpipe_model)
}
}
