% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\docType{data}
\name{fairy}
\alias{fairy}
\title{Fifty H.C. Andersen’s fairy tales}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 50 rows and 4 columns.}
\usage{
data(fairy)
}
\description{
First fifty fairy tales from hcandersenr::hcandersen_en.
}
\examples{
\dontrun{
data(fairy)
fairy_rawscore <- calculate_textplex(fairy$text)
}
}
\keyword{datasets}
