% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\docType{data}
\name{fairy_rawscore}
\alias{fairy_rawscore}
\title{Raw scores extracted from the fairy dataset}
\format{An object of class \code{data.frame} with 50 rows and 10 columns.}
\usage{
data(fairy)
}
\description{
Raw scores extracted from the fairy dataset using calculate_textplex
}
\examples{
data(fairy_rawscore)
fit <- fit_two_factor_model(fairy_rawscore)
}
\keyword{datasets}
