% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\docType{data}
\name{sotu_rawscore}
\alias{sotu_rawscore}
\title{Raw scores extracted from the state of the union addresses}
\format{An object of class \code{data.frame} with 236 rows and 9 columns.}
\usage{
data(sotu_rawscore)
}
\description{
Raw scores extracted from the state of the union addresses
}
\examples{
data(sotu_rawscore)
fit <- fit_two_factor_model(sotu_rawscore)
}
\keyword{datasets}
