% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{calculate_textplex}
\alias{calculate_textplex}
\title{Calculate the raw scores}
\usage{
calculate_textplex(
  input_text,
  parsed_text = NULL,
  mtld_threshold = 0.72,
  mtld_min_ntoken = 10,
  parallel = TRUE,
  .progress = TRUE,
  plan = "multiprocess"
)
}
\arguments{
\item{input_text}{vector of characters.}

\item{parsed_text}{a list of dataframes. Must be parsed with spacyr::spacy_parse(pos = TRUE, dependency = TRUE, tag = TRUE)}

\item{mtld_threshold}{Threshold for the calculation of MTLD score.}

\item{mtld_min_ntoken}{Threhold for minimal length of a viable sentence.}

\item{parallel}{logical, whether or not to use parallel computation.}

\item{plan}{character, please refer to ?future::plan for other options.}

\item{.progrss_bar}{logical, whether or not to display progress bar. Only possible when parallel = TRUE.}
}
\value{
tibble of raw scores
}
\description{
This function takes a vector of characters and calculate the raw scores required for the algorithm.
}
\examples{
\donttest{
require(spacyr)
data(fairy)
spacy_initialize(model = 'en')
fairy_rawscore <- calculate_textplex(fairy$text)
}
}
