% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{fit_two_factor_model}
\alias{fit_two_factor_model}
\title{Fit a two-factor model as per Tolochko and Boomgaarden}
\usage{
fit_two_factor_model(raw_scores)
}
\arguments{
\item{raw_scores}{raw_scores extracted by calculate_textplex()}
}
\value{
a psych::principal fit.
}
\description{
This function fits a two-factor model of text complexity as per Tolochko and Boomgaarden (2019).
}
\examples{

}
